+++ src/contexts/firebase.tsx
@@ -1,12 +1,13 @@
 import React, { createContext, useContext, useEffect, useState } from "react";
-import Constants, { ExecutionEnvironment } from "expo-constants";
 import uuid from "react-native-uuid";
+import Crashlytics from "@react-native-firebase/crashlytics"
 
 import * as Analytics from "expo-firebase-analytics";
 import { useAuth } from "./auth";
 
 type FirebaseContextProps = {
   analytics: typeof Analytics;
+  crashlytics: typeof Crashlytics | null;
 };
 
 const FirebaseContext = createContext<FirebaseContextProps>(
@@ -15,6 +16,7 @@ const FirebaseContext = createContext<FirebaseContextProps>(
 
 const FirebaseProvider: React.FC = ({ children }) => {
   const [analytics, setAnalytics] = useState<typeof Analytics>(Analytics);
+  const [crashlytics, setCrashlytics] = useState<typeof Crashlytics | null>(null);
   const { user } = useAuth();
 
   useEffect(() => {
@@ -24,16 +26,28 @@ const FirebaseProvider: React.FC = ({ children }) => {
       }
 
       Analytics.setClientId(uuid.v4().toString());
-      // await Analytics.setDebugModeEnabled(__DEV__);
 
       setAnalytics(analytics);
     })();
   }, []);
 
+  useEffect(() => {
+    (async () => {
+      if (user) {
+        await Crashlytics().setUserId(String(user?.id))
+        await Crashlytics().setAttributes(user as {})
+      }
+  
+      await Crashlytics().setCrashlyticsCollectionEnabled(!__DEV__)
+      setCrashlytics(Crashlytics)
+    })()
+  }, [])
+
   return (
     <FirebaseContext.Provider
       value={{
         analytics,
+        crashlytics
       }}
     >
       {children}